Better card handling
There's a new script, `script/importCards.js`, for importing cards from the PYX SQL dump. The result of this is in the `config/cards` directory. The stock cards for the irc-cah game are gone, replaced with the PYX cards.

Additionally, the card loading logic has been changed to reference the config file to find the available decks, instead of requiring the user to edit the game's code. A future direction is to support adding and removing decks at runtime, and for querying the bot about the loaded decks.