Add a script for importing cards, add the Fifth Expansion
The script accepts a list of cards on standard input and takes
two commandline arguments: --type Question|Answer and --name
deckName. The question cards use %s to indicate blanks (just
like the existing decks), and the script calculates the pick
and draw values based on the number of %s.

Remaining to do: import the 10 days or whatever of Kwanza cards.